+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 
+## [0.2.0] - 2023-07-09
+### Added
+- `Entry` API for `BTreeMap`
+- `serde` feature for deserialization of `BTreeSet` and `BTreeMap`
+
 ## [0.1.0] - 2023-07-04
 ### Added
 - `BTreeSet`